Enhance documentation and components with new features. Updated `documentation.md` to include detailed descriptions of new components: `NoteEditor`, `NotesSidebar`, and `SearchBar`. Improved `README.md` to reflect enhanced note management features, keyboard shortcuts, and responsive design. Added custom hooks for theme management and keyboard shortcuts to streamline functionality in the application.
