+++ documentation.md
@@ -6,6 +6,13 @@
 taskmark/
 ├── src/
 │   ├── components/         # Reusable UI components
+│   │   ├── NoteEditor.jsx  # Note editing component
+│   │   ├── NotesSidebar.jsx # Sidebar component
+│   │   └── SearchBar.jsx   # Search component
+│   ├── hooks/             # Custom React hooks
+│   │   ├── useTheme.js    # Theme management hook
+│   │   ├── useKeyboardShortcuts.js # Keyboard shortcuts hook
+│   │   └── useNoteManagement.js # Note management hook
 │   ├── App.jsx            # Main application component
 │   ├── NotesContext.jsx   # Notes state management
 │   ├── RecycleBinModal.jsx # Recycle bin modal component
@@ -16,19 +23,14 @@ taskmark/
 ## Core Components
 
 ### App.jsx
-The main application component that handles:
+The main application component that orchestrates the entire application.
+
+Key features:
 - Theme management
 - Keyboard shortcuts
 - Layout structure
 - Modal states (search, recycle bin)
 - Mobile responsiveness
-
-Key features:
-- Theme cycling between pastel, light, and dark modes
-- Keyboard shortcuts for common actions
-- Responsive layout with mobile sidebar
-- Search functionality
-- Recycle bin management
 
 ### NotesContext.jsx
 State management for notes using React Context API.
@@ -40,15 +42,50 @@ Key functions:
 - `permanentlyDeleteNote`: Permanently removes a note
 - `undoDelete`: Restores the last deleted note
 
+### NoteEditor.jsx
+Component for editing individual notes.
+
+Features:
+- Title and content editing
+- Block-based text editing
+- Tag management
+- Word and line count statistics
+- Last updated timestamp
+- Keyboard shortcuts support
+- Auto-save functionality
+
+### NotesSidebar.jsx
+Component for displaying and managing the list of notes.
+
+Features:
+- List of all notes
+- Note selection
+- Search functionality
+- Mobile responsiveness
+- Note creation button
+- Last updated time display
+
+### SearchBar.jsx
+Component for searching through notes.
+
+Features:
+- Real-time search
+- Search through titles, content, tags, and dates
+- Keyboard navigation
+- Search suggestions with preview
+- Last updated time display
+- Tag display in results
+
 ### RecycleBinModal.jsx
 Modal component for managing deleted notes.
 
 Features:
-- Lists all deleted notes with deletion timestamps
+- Lists all deleted notes
 - Restore functionality
 - Permanent deletion option
 - Keyboard support (Esc to close)
 - Responsive design
+- Deletion timestamp display
 
 ### Notification.jsx
 Toast-style notification component for user feedback.
@@ -58,134 +95,97 @@ Features:
 - Undo functionality
 - Customizable duration
 - Clean animations
+- Theme-aware styling
 
-## Keyboard Shortcuts
+## Custom Hooks
 
-| Shortcut | Action |
-|----------|---------|
-| `Alt + N` | Create new note |
-| `Alt + S` | Open search |
-| `Alt + Delete` | Delete current note |
-| `Esc` | Close search/recycle bin |
-| `Enter` | Select note from search |
+### useTheme
+Manages application theming.
+
+Features:
+- Theme cycling (Pastel, Light, Dark)
+- Theme persistence in localStorage
+- Meta theme color updates
+- Theme-aware styling
+
+### useKeyboardShortcuts
+Manages keyboard shortcuts throughout the application.
+
+Shortcuts:
+- `Alt + N`: Create new note
+- `Alt + S`: Open search
+- `Alt + T`: Add tag
+- `Alt + Delete`: Delete current note
+- `Esc`: Close search/recycle bin
+- `Enter`: Select note from search
+
+### useNoteManagement
+Manages note-related state and operations.
+
+Features:
+- Note creation
+- Note updates
+- Note deletion
+- Note selection
+- Note sorting
+- Search functionality
 
 ## Theme System
 
-The application uses CSS variables for theming with three built-in themes:
+The application uses CSS variables for theming, with three built-in themes:
+- Pastel (default)
+- Light
+- Dark
 
-### Pastel Theme (Default)
-```css
---bg-primary: #1A3636;
---bg-secondary: #40534C;
---bg-tertiary: #2E4B4B;
---text-primary: #D6BD98;
---text-secondary: #a9997c;
---accent: #D6BD98;
-```
-
-### Light Theme
-```css
---bg-primary: #ffffff;
---bg-secondary: #f5f5f5;
---bg-tertiary: #f9f9f9;
---text-primary: #1a1a1a;
---text-secondary: #555555;
---accent: #0066cc;
-```
-
-### Dark Theme
-```css
---bg-primary: #121212;
---bg-secondary: #1e1e1e;
---bg-tertiary: #252525;
---text-primary: #e0e0e0;
---text-secondary: #a0a0a0;
---accent: #bb86fc;
-```
+Each theme defines:
+- Background colors
+- Text colors
+- Border colors
+- Accent colors
+- Hover states
 
 ## State Management
 
-The application uses React Context for state management with the following key states:
+The application uses React Context for state management:
+- Notes state (active and deleted notes)
+- Theme state
+- UI state (modals, search, etc.)
 
-### Notes State
-- `notes`: Array of active notes
-- `deletedNotes`: Array of notes in recycle bin
-- `lastDeletedNote`: Reference to the most recently deleted note
-- `loading`: Loading state for initial data fetch
+## Data Persistence
 
-### UI State
-- `theme`: Current theme selection
-- `isMobileSidebarOpen`: Mobile sidebar visibility
-- `isSearchOpen`: Search modal visibility
-- `isRecycleBinOpen`: Recycle bin modal visibility
-- `searchQuery`: Current search input
+Notes are persisted in:
+- localStorage for active notes
+- localStorage for deleted notes
+- localStorage for theme preference
 
-## Event Handling
+## Performance Considerations
 
-### Keyboard Events
-- Global keyboard event listener in App.jsx
-- Handles shortcuts for note creation, deletion, and search
-- Modal closing with Escape key
-- Search result navigation
+- Memoized components and callbacks
+- Efficient state updates
+- Lazy loading of components
+- Optimized re-renders
+- Debounced search
+- Efficient DOM updates
 
-### Click Events
-- Note selection
-- Theme toggling
-- Modal opening/closing
-- Note restoration/deletion
+## Accessibility
 
-## Styling
-
-The application uses:
-- Tailwind CSS for utility classes
-- CSS variables for theming
-- Custom CSS for specific components
-- Responsive design patterns
-- Font: Fira Code (monospace)
-
-## Best Practices
-
-1. **Component Organization**
-   - Reusable components in `components/` directory
-   - Main application logic in `App.jsx`
-   - State management in `NotesContext.jsx`
-
-2. **State Management**
-   - Use Context API for global state
-   - Local state for component-specific data
-   - Proper state updates with immutable patterns
-
-3. **Performance**
-   - Memoization where appropriate
-   - Efficient event handling
-   - Optimized re-renders
-
-4. **Accessibility**
-   - Keyboard navigation support
-   - ARIA labels where needed
-   - Semantic HTML structure
-
-5. **Code Style**
-   - Consistent naming conventions
-   - Component-based architecture
-   - Clear separation of concerns
+- Keyboard navigation
+- ARIA labels
+- Semantic HTML
+- Focus management
+- Screen reader support
+- Color contrast compliance
 
 ## Future Improvements
 
-1. **Features**
-   - Note categories/folders
-   - Rich text editing
-   - Note sharing
-   - Export/import functionality
-
-2. **Technical**
-   - Unit testing
-   - E2E testing
-   - Performance optimization
-   - Offline support
-
-3. **UI/UX**
-   - More theme options
-   - Customizable shortcuts
-   - Enhanced mobile experience
-   - Better accessibility 
+Potential areas for enhancement:
+- Rich text editing
+- Note categories
+- Export/Import functionality
+- Cloud sync
+- Collaborative editing
+- Unit testing
+- Performance optimization
+- Offline support
+- Markdown support
+- Note sharing
